deepeval_demo/requirements.txt
deepeval>=2.0.0
python-dotenv>=1.0.0

